# docs/openapi-nvin.yaml
openapi: 3.0.3
info:
  title: Field Commander — NVIN Integration API
  description: |
    REST API for integrating Field Commander mobile agents with the
    National Vehicle Intelligence Network (NVIN / InsightLPR / Odasi).
  version: 1.0.0
  contact:
    name: NVIN Integration Team
servers:
  - url: https://api.nvin.network/v1
    description: Production
  - url: http://localhost:3001/v1
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /dta/payout:
    post:
      summary: Trigger DTA payout for verified vehicle recovery
      description: |
        Calls CitizenLedgerDTA on Polygon to execute ETH + TCR payout
        for a field agent who recovered a listed vehicle.
      operationId: requestDTAPayout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DTAPayoutRequest'
      responses:
        '200':
          description: Payout executed on-chain
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DTAPayoutResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          description: Confidence below minimum threshold (85)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DTAPayoutRequest:
      type: object
      required: [plate_hash, agent_address, evidence_hash, confidence]
      properties:
        plate_hash:     { type: string, description: 'SHA-256 of plate text (hex)' }
        agent_address:  { type: string, description: 'Ethereum wallet address' }
        evidence_hash:  { type: string, description: 'SHA-256 of evidence bundle' }
        confidence:     { type: number, minimum: 85, maximum: 100 }
    DTAPayoutResponse:
      type: object
      properties:
        transaction_hash: { type: string }
        payout_amount:    { type: string, description: 'ETH amount as decimal string' }
        tcr_reward:       { type: string, description: 'TCR token amount' }
        block_number:     { type: integer }
...
    Error:
      type: object
      properties:
        error: { type: string }
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...